refactor: small improvements to copy() call in s3_v2

+++ registry/storage/driver/s3-aws/v1/s3.go
@@ -77,7 +77,7 @@ type driver struct {
 	Encrypt                     bool
 	KeyID                       string
 	MultipartCopyChunkSize      int64
-	MultipartCopyMaxConcurrency int64
+	MultipartCopyMaxConcurrency int
 	MultipartCopyThresholdSize  int64
 	RootDirectory               string
 	StorageClass                string
